+++ source/Node/Voxelization/VoxelizationConfig.cpp
@@ -238,7 +238,7 @@ QCStatus_e VoxelizationConfig::ParseStaticConfig( DataTree &dt, std::string &err
     if ( QC_STATUS_OK == ret )
     {
         config.nodeId.name = dt.Get<std::string>( "name", "" );
-        config.nodeId.id = dt.Get<uint32_t>( "id", UINT32_MAX );
+        config.nodeId.id = dt.Get<uint8_t>( "id", UINT8_MAX );
         config.voxelConfig.processor = dt.GetProcessorType( "processorType", QC_PROCESSOR_HTP0 );
         config.voxelConfig.coreId = dt.Get<uint32_t>( "coreId", 0 );
         config.voxelConfig.pillarXSize = dt.Get<float>( "Xsize", 0.0f );
